call tar with -o flag
When tar is run as root it attempts to preserve user & group ownership by
default. If the uid and gid don't exist this can cause tar to fail. A good
example is Number-Compare 0.03, which have uid 831580115 and gid
755412454. Attempting to untar this as root will fail.

While normally we don't expect to be run as root, this is quite common when
running Docker, which in turn is used by many CI environments, including
CircleCI.