Fix unpacking when using OpenBSD tar
By default OpenBSD ships with its own version of tar(1) which is
neither bsdtar, nor GNU. This version does not use -o parameter in the
same way other two do. Instead OpenBSD tar always work as if -o is
enabled, even for root user, and has an option (-p) to disable this
behavior.

Currently this is not taken into account. On OpenBSD if executed with
-o parameter, tar prints a warning and exits with an error code.

To accommodate for that, extend tar(1) version detection to
specifically include GNU (which can also be present on OpenBSD
system), BSD, and OpenBSD versions of the tar(1) program.

There is no easy way to detect OpenBSD tar executable, as it does not
provide --version output. When OS is detected to be OpenBSD, no
additional switch is used for tar invocation, otherwise -o is kept as
before.